+++ format/xm/system/system.go
@@ -27,7 +27,7 @@ const (
 	C4SlideFines          = C4Note * SlideFinesPerNote
 )
 
-var semitonePeriodTable = [...]float32{27392, 25856, 24384, 23040, 21696, 20480, 19328, 18240, 17216, 16256, 15360, 14496}
+var semitonePeriodTable = [...]uint16{27392, 25856, 24384, 23040, 21696, 20480, 19328, 18240, 17216, 16256, 15360, 14496}
 
 var XMSystem system.ClockableSystem = system.ClockedSystem{
 	MaxPastNotesPerChannel: 0,
